Lançamento de exceção ao remover livros associados a autores

+++ src/main/java/servico/Servico.java
@@ -1,44 +1,46 @@
 package servico;
 
 import entidades.EntidadeNegocio;
+import exception.NegocioException;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.PersistenceContextType;
 import javax.persistence.TypedQuery;
 
-public abstract class Servico<T extends EntidadeNegocio> {
-
+public abstract class Servico<T extends EntidadeNegocio>
+{
     @PersistenceContext(unitName = "rater-book", type = PersistenceContextType.TRANSACTION)
     protected EntityManager em;
 
-    public void salvar(T entidadeNegocio) {
-
+    public void salvar(T entidadeNegocio)
+    {
         em.persist(entidadeNegocio);
-
     }
 
-    public void alterar(T entidadeNegocio) {
-
+    public void alterar(T entidadeNegocio)
+    {
         em.merge(entidadeNegocio);
-
     }
 
-    public void remover(T entidadeNegocio) {
-
-        entidadeNegocio = em.find(getClasseEntidade(), entidadeNegocio.getChavePrimaria());
-        em.remove(entidadeNegocio);
-
+    public void remover(T entidadeNegocio) throws NegocioException
+    {
+        if (entidadeNegocio.associado() == false)
+        {
+            entidadeNegocio = em.find(getClasseEntidade(), entidadeNegocio.getChavePrimaria());
+            em.remove(entidadeNegocio);
+        }
+        else throw new NegocioException(NegocioException.ENTIDADE_ASSOCIADA);
     }
 
-    public List<T> listarTodos() {
+    public List<T> listarTodos()
+    {
         StringBuilder jpql = new StringBuilder();
         jpql.append("select e ");
         jpql.append(" from ");
         jpql.append(getClasseEntidade().getSimpleName());
         jpql.append(" as e ");
-        TypedQuery<T> query = em.createQuery(jpql.toString(), 
-                getClasseEntidade());
+        TypedQuery<T> query = em.createQuery(jpql.toString(), getClasseEntidade());
         return query.getResultList();
     }
